fix : Library

+++ app/src/main/java/com/msb/purrytify/data/local/dao/SongDao.kt
@@ -16,13 +16,13 @@ interface SongDao {
     suspend fun delete(song: Song)
 
     @Query("SELECT * FROM songs WHERE ownerId = :ownerId ORDER BY addedAt DESC")
-    fun getAllSongs(ownerId: Int): Flow<List<Song>>
+    fun getAllSongs(ownerId: Long): Flow<List<Song>>
 
     @Query("SELECT * FROM songs WHERE isLiked = 1 AND ownerId = :ownerId ORDER BY addedAt DESC")
-    fun getLikedSongs(ownerId: Int): Flow<List<Song>>
+    fun getLikedSongs(ownerId: Long): Flow<List<Song>>
 
     @Query("SELECT * FROM songs WHERE lastPlayedAt IS NOT NULL AND ownerId = :ownerId ORDER BY lastPlayedAt DESC LIMIT 10")
-    fun getRecentlyPlayedSongs(ownerId: Int): Flow<List<Song>>
+    fun getRecentlyPlayedSongs(ownerId: Long): Flow<List<Song>>
 
     @Query("SELECT * FROM songs WHERE id = :songId")
     suspend fun getSongById(songId: Long): Song?
@@ -34,14 +34,14 @@ interface SongDao {
     suspend fun updateLastPlayedAt(songId: Long, timestamp: Long)
 
     @Query("SELECT COUNT(*) FROM songs WHERE ownerId = :ownerId")
-    fun getSongCount(ownerId: Int): Flow<Int>
+    fun getSongCount(ownerId: Long): Flow<Int>
 
     @Query("SELECT COUNT(*) FROM songs WHERE isLiked = 1 AND ownerId = :ownerId")
-    fun getLikedSongCount(ownerId: Int): Flow<Int>
+    fun getLikedSongCount(ownerId: Long): Flow<Int>
 
     @Query("SELECT COUNT(*) FROM songs WHERE lastPlayedAt IS NOT NULL AND ownerId = :ownerId")
-    fun getListenedSongCount(ownerId: Int): Flow<Int>
+    fun getListenedSongCount(ownerId: Long): Flow<Int>
 
     @Query("SELECT * FROM songs WHERE ownerId = :ownerId ORDER BY addedAt DESC LIMIT 10")
-    fun getNewSongs(ownerId: Int): Flow<List<Song>>
+    fun getNewSongs(ownerId: Long): Flow<List<Song>>
 }
